artifactcache: return overhead required to store an artifact
Track amount of bytes required to commit or pull an artifact and return
this data on pull() and commit().

The commit() return value is more accurate than computing the size of
the artifact staging directory, as it takes deduplication into account.

This will be used in a following commit to dynamically update the cache
size.