Fix ostree repository mirroring
Ostree mirrors were not sharing the same local repository, so it was
impossible the request refs from the right local repository when data
was fetched from a mirror rather than upstream.

Instead of having several repository with one remote each, we now
have one repository with several remotes.

This fixes #538.