Invalidate reverse dependencies to scheduled element with workspace
Issues:

Keys would change from one build to another. So building a second time
an element depending a workspaced element, would fail to find the
artifact that was already built.

Artifact cache would be populated with artifacts with the wrong
key. This is a serious issue were an artifact that was built with a
workspaced dependency would not rebuild after closing the workspace.

Solution:

Elements depending on another element with workspace which has
been scheduled should not have any keys until we have managed to
built the workspaced element.

Fixes #461.