dev-requirements.txt: Pin versions of pytest and pylint
These packages should only be updated as a deliberate change and the
code should be tested with the new versions to ensure that the test
still pass.

Any future updates should also happen in sync with the
buildstream-docker-images repository so that the testsuite base images
also enforce the same restrictions for the development requirements.

Replaces https://gitlab.com/BuildStream/buildstream/merge_requests/684.