README.rst: Removing the badges
Now that we only run pre-merge CI, it makes no sense to show a badge for
the coverage and pipeline status of the last run of master.

It would make sense to show a badge for the last pipeline which
was run for the last merge request that was merged to master, but
I don't see any feature to get that information.