.gitlab-ci.yml: Verify that source tarball installs correctly
Add a check to verify that the source distribution tarball that we
generate installs corrctly. It is useful to do this check as this is the
same tarball will be used when installing BuildStream via PyPI.

This check also happens indirectly in the coverage job where we install
BuildStream but this check makes it more explicit and will also cause
the CI to fail early if the tarball is not correct.