Refactor casserver for better coverage
Previously we had the logic for setting up the ports on the server in
server_main. This makes for messy testing of this functionality. As
such, this setup logic has been moved to a public function which can be
tested easily.

This commit does not change any exists test, but modifies the testutil
for creating a cache server.