artifactcache: dynamically update size all time
Change pull jobs to dynamically update the cache size, as build jobs
were already doing it.

The pathways between pull and build jobs are similiar and the code can
be updated appropriately.

As the cache size is always dynamically updated, we no longer need to
run the CacheSize job.

Fixes #573.