_frontend/status.py: Completely remove the blessings dependency from BuildStream
This actually improves reliability of the status bar because we
now disable it completely in the case that not all of the terminal
escape sequences are supported on the given terminal.

This replaces the few functions we were using, to move the cursor
up one line, move it to the beginning of the line, and to clear a
line, with low level functions provided by the curses module in
the standard library.

This change makes it easier for downstream distro package maintainers
to package BuildStream, particularly on Fedora.

Asides from changing _frontend/status.py, this commit includes the
following changes:

  * _frontend/app.py: Use python isatty() function to determine if
                      we are connected to a tty, instead of relying
		      on blessings.

  * setup.py: Remove the dependency on blessings.