_platform/linux.py: Refactor checks for sandboxing
To better report issues in the absence of a suitable bwrap, or the
FUSE devices, this refactors the checks for sandboxing in the Linux
platform to cover the various possibilities.  The reasons are then
collated and passed to the dummy sandbox for later reporting to the
user if a local build is attempted.

Signed-off-by: Daniel Silverstone <daniel.silverstone@codethink.co.uk>