Upstream freedesktop-sdk strip rules
Adds a modified version of the freedesktop-sdk `strip-binaries` to
buildstream default `project.conf`. We implemented a few changes to
the current default, but the main change is to fix a bug due to
permissions.

The modifications I made to the freedesktop-sdk commands as they are
at present are as follows:
* Remove debugedit rules to include source debugging
* Add a switch to allow failing on error, rather than this being
  the default behaviour.
  These changes mean that we should now have a sensible default for
  a wide variety of users.