utils.py: Disable lint warning about popen
We use `preexec_fn` in our use of Popen.  Since there are threads
in the codebase, this is not necessarily safe.  However we go to
great lengths to ensure that the main process doesn't spawn threads
(they're used in gRPC a lot) and so it should be safe for our use.
As such, we disable the lint here.

Signed-off-by: Daniel Silverstone <daniel.silverstone@codethink.co.uk>