element: Add BST_STAGE_INTEGRATES Element flag
This commit introduces BST_STAGE_INTEGRATES as a compatibility flag,
so Elements can set it to false
to signal that integrate_dependency_artifacts should be called.

This will be needed to implement staging multiple elements,
since the integration commands for all elements need to be called
after all dependencies are staged.