elements: Port build and script kinds to BST_STAGE_INTEGRATES
This is mostly just marking which elements work as expected
from changes to their base class.

Junction and Filter elements expect overriding stage to be sufficient
to prevent it doing anything in those contexts,
and since we're not intending to deprecate BST_STAGE_INTEGRATES elements
there's no harm in them continuing to do so
when they aren't expected to work in contexts we require
BST_STAGE_INTEGRATES to be False.