Keep the Bst2html output up to date.
This patch allows bst2html to run when the rest of the docs are rebuilt
and remove the old stale cached files.

This patch also makes it easier to maintain the .run files by having
bst2html have a less aggressive caching policy that is now based on
mtimes mtimes.