tests/.../workspace: assert --assume-yes is needed
Add tests to ensure that these commands will fail if we are not
interactive:

- bst workspace close --remove-dir
- bst workspace reset

They should be invoked with '--assume-yes' if
'prompt.really-workspace-*' options are not configured.

Also make these code paths testable by raising an AppError instead of
sys.exit(-1).

This also tests the 'prompt.really-workspace-*' options, they were not
testable before the '--assume-yes' breaking change to non-interactive
behaviour.

Note that interactive behaviour is still not covered and must be tested
manually.