man/: update with changes since Apr 2018
It seems that the regenerated man pages haven't been committed for some
time. Add them all in one lump now.

Had to manually strip the version number from the generated files.

Had to manually fix the NAME field of bst.1 to not be 'bst - None'.

Oddly had to invoke click-man directly in order to get the changes to
be applied, the usual setup.py route mentioned in contributing.rst
didn't work.

Eyeballed the result of invoking man on each of the commands, seems
good.