plugins/sources/git.py: Avoid downloading unused submodules
Currently we have configuration in place to disable use of
submodules, with the `checkout-submodules` git plugin option
and the individual `checkout` options for each submodule, but
these unused submodules are still downloaded at fetch time.

This patch fixes the plugin to just completely ignore the
submodules which are configured to be unused.

This was previously fixed in the master branch by !996, and
this patch backports the relevant fix to the 1.2 branch,
fixing issue #804.