cli: Add artifact checkout subcommand
This implements a superset of the features of `bst checkout`,
but uses a --directory option rather than a location argument,
and --tar takes a location instead of being a flag.

If multiple element targets or any artifact refs are passed
then it does not support creating a tarball, running integration commands
or including dependencies of the targets.