tmpdir: add tmpdir to context for CASRemote
As CASRemote is used in other places such as SandboxRemote it makes sense for
there to be a tmpdir it uses. This dir is passed to CASRemote when initialized.

This is currently in the artifactdir, but should be moved when artifactdir and
builddir options are deprecated in favour of one top level directory containing
the two.

Part of #802